[GTK][WPE][Skia] Visiting wpewebkit.org with Skia/CPU rendering fires assertion https://bugs.webkit.org/show_bug.cgi?id=285768
Reviewed by Adrian Perez de Castro.

drawFilteredImageBuffer() tries to determine if it needs to copy the
ImageBuffer backend of the accelerated image, in case of cross-thread
usage (e.g. re-wrapping an existing GPU resource in another thread).
This erroneously ran also for unaccelerated rendering, firing an
assertion that the SkiaGLContext is not available, which is correct. We
shouldn't attempt to run the copy-ImageBuffer-backend logic during
unaccelerated rendering -- fix that.

* Source/WebCore/platform/graphics/skia/GraphicsContextSkia.cpp:
(WebCore::GraphicsContextSkia::drawFilteredImageBuffer):

Canonical link: https://commits.webkit.org/288778@main