[iOS] TestWebKitAPI.AppHighlights.AppHighlightCreateAndRestoreAndScroll is timing out https://bugs.webkit.org/show_bug.cgi?id=285882 rdar://141462057
Reviewed by Abrar Rahman Protyasha and Megan Gardner.

This is the same test development issue as 287822@main, and as such the commit
message below is copied verbatim.

Since TestWebKitAPI is not a real `UIApplication`, UIKit gets into a state
where its own update cycle is enabled (which is not runtime configurable), but
not started (which is done in `UIApplicationMain()`).

This means that `UIAnimation`-driven animations (like scroll view animations)
will never reach completion, and `scrollViewDidEndScrollingAnimation:` will
never be called.

Fix by leveraging the fact that UIKit's update cycle is not used for
`UIAnimation`s if `-[UIScreen _isEmbeddedScreen]` is false. This will result in
UIKit using a `CADisplayLink` to drive animations, and ensures completion is
reached without needing to be a `UIApplication`.

* Tools/TestWebKitAPI/Tests/WebKitCocoa/WKAppHighlights.mm:
(TestWebKitAPI::TEST(AppHighlights, AppHighlightCreateAndRestoreAndScroll)):

Canonical link: https://commits.webkit.org/288848@main