Make UIScriptContext::asyncTaskComplete more robust https://bugs.webkit.org/show_bug.cgi?id=285930 rdar://142903461
Reviewed by Pascoe.

I have crash logs that indicate that it can sometimes be called with a key of 0,
and it can sometimes be called with a key that returns a null callback, possibly
as state leftover from a previous test run that initiated an async task that was
not cancelled.  When this happens, ignore the call instead of crashing.

* Tools/TestRunnerShared/UIScriptContext/UIScriptContext.cpp:
(UIScriptContext::asyncTaskComplete):

Canonical link: https://commits.webkit.org/288886@main