Early return instead of terminating web process when navigating destroyed frames https://bugs.webkit.org/show_bug.cgi?id=285925 rdar://142902541
Reviewed by Pascoe.

This is similar to 287896@main but also fixing the MESSAGE_CHECK_COMPLETION calls that check for a frame's existence.
There is a race condition introduced with site isolation that the UI process can destroy a WebFrameProxy and
notify each process that the frame has been destroyed, but before a process receives that message it sends
the UI process a message indicating that the frame is navigating.  When this happens, it is correct to do
nothing, or to complete the request to indicate that we should do nothing.

* Source/WebKit/UIProcess/WebPageProxy.cpp:
(WebKit::WebPageProxy::decidePolicyForNavigationActionAsyncShared):
(WebKit::WebPageProxy::decidePolicyForNavigationActionSyncShared):
(WebKit::WebPageProxy::decidePolicyForResponseShared):

Canonical link: https://commits.webkit.org/288887@main