[iOS] Base writing direction should update to reflect active input mode after inserting a newline https://bugs.webkit.org/show_bug.cgi?id=285960 rdar://142411271
Reviewed by Abrar Rahman Protyasha.

Add a heuristic to automatically set the base writing direction to match the character direction of
the user's active input mode on iOS, after inserting a new paragraph. To achieve this, we use the
content view's `textInputMode.primaryLanguage` to compute `directionFromCurrentInputMode`, and send
it through `InsertTextOptions` to the web process. If we're inserting only a newline and the
selection is a caret, then we adjust the base writing direction to match the incoming
`directionFromCurrentInputMode`.

* LayoutTests/editing/input/ios/update-writing-direction-after-inserting-newline-expected.txt: Added.
* LayoutTests/editing/input/ios/update-writing-direction-after-inserting-newline.html: Added.

Add a layout test to exercise this change.

* Source/WebKit/Shared/Cocoa/InsertTextOptions.h:
* Source/WebKit/Shared/Cocoa/InsertTextOptions.serialization.in:

Add an optional `directionFromCurrentInputMode` member to `InsertTextOptions`.

* Source/WebKit/UIProcess/ios/WKContentViewInteraction.mm:
(-[WKContentView insertText:]):
* Source/WebKit/WebProcess/WebPage/WebPage.cpp:
(WebKit::WebPage::insertTextAsync):

See description above for more details.

Canonical link: https://commits.webkit.org/288911@main