REGRESSION:(288166@main) [ iOS ] 2x TestWebKitAPI.SiteIsolation.* (api-tests) are constant asserts https://bugs.webkit.org/show_bug.cgi?id=285554 rdar://142500396
Reviewed by Simon Fraser.

For cases where a scrolling state tree is added to the list of hosted subtrees needing connection (parent tree hasn’t been added
to scrolling tree yet) -> a corresponding remote layer tree node on said tree is removed -> the parent tree is added to the tree ->
we try to connect the node that had its remote layer tree node removed and crash when trying to access the scrollView (which is
normally held on the remote layer tree node). This is not an invalid configuration to be in since the scrolling node will be properly
be removed once we get through all of the pending subtrees needing connection, so we only need to address the crash itself.

* Source/WebKit/UIProcess/RemoteLayerTree/ios/ScrollingTreeScrollingNodeDelegateIOS.mm:
(WebKit::ScrollingTreeScrollingNodeDelegateIOS::scrollView const):

Canonical link: https://commits.webkit.org/288914@main