[cocoa] Cache opt-in partitioned cookie enablement in Web Process and use it when creating cookie cache https://bugs.webkit.org/show_bug.cgi?id=286016 rdar://142985374
Reviewed by Alex Christensen.

On some platforms where the WebCookieCache is supported, the cache can be
destroyed at any time, such as when the process is suspended. We need to
remember if opt-in cookie partitioning was enabled so we can enable it in the
cache when that is recreated. This patch caches the state in the WebProcess and
uses the singleton when instantiating a new in-memory cache on Cocoa platforms.

* Source/WebKit/WebProcess/WebPage/Cocoa/WebCookieCacheCocoa.mm:
(WebKit::WebCookieCache::inMemoryStorageSession):
* Source/WebKit/WebProcess/WebProcess.cpp:
(WebKit::WebProcess::setOptInCookiePartitioningEnabled):
* Source/WebKit/WebProcess/WebProcess.h:

Canonical link: https://commits.webkit.org/288981@main