REGRESSION (282580@main): font-size: calc(). using rem and em prevents text from scaling up or down https://bugs.webkit.org/show_bug.cgi?id=285914
Reviewed by Simon Fraser.

Style::computeNonCalcLengthDouble doesn't apply zoom when processing the font-size
property, so unconditionally trying to remove zoom in the calc code path is incorrect.

This adds a special case for font-size, which fixes the issue, but a better fix would
be to refactor Style::computeNonCalcLengthDouble to explicitly track when zoom is applied.

* LayoutTests/fast/css/font-size-calc-with-em-zoom-expected.txt: Added.
* LayoutTests/fast/css/font-size-calc-with-em-zoom.html: Added.
* Source/WebCore/css/calc/CSSCalcTree+Simplification.cpp:
* Source/WebCore/style/values/primitives/StyleLengthResolution.cpp:
* Source/WebCore/style/values/primitives/StyleLengthResolution.h:

Canonical link: https://commits.webkit.org/289030@main