AX: [Site Isolation] Sites with multiple iframes don't initialize remote frames https://bugs.webkit.org/show_bug.cgi?id=285951 rdar://142913087
Reviewed by Chris Fleizach.

The LocalFrame is not always the first descendant of the mainFrame's tree. This
is the case, for example, when there are multple remote frames on one page.

We should instead use find the local frame that matches the web page object's
frame identifier.

* Source/WebKit/WebProcess/WebPage/mac/WKAccessibilityWebPageObjectBase.mm:
(-[WKAccessibilityWebPageObjectBase remoteLocalFrame]):

Canonical link: https://commits.webkit.org/289066@main