NetworkProcess does not always terminate properly on macOS https://bugs.webkit.org/show_bug.cgi?id=285947 rdar://142838090
Reviewed by Per Arne Vollan.

NetworkProcess does not always terminate on macOS. The reason for this is that NetworkProcess can
suspend in the middle of termination:

1. The NetworkProcessProxy destructor drops "m_backgroundActivityToPreventSuspension", which allows
NetworkProcess to suspend.
2. Then the AuxiliaryProcessProxy destructor runs and starts termination by invalidating via
IPC::Connection::invalidate.
3. The remote side of the connection invalidation in NetworkProcess (NetworkProcess::didClose) might
never run if the process is already suspended, which then leaves the process in a permanently
suspended state.

To fix this, as part of AuxiliaryProcessProxy destruction, we now start a watchdog timer which
allows the process to keep running while termination is active, and we kill the process when the
timer expires. This used to be part of XPCConnectionTerminationWatchdog, but I removed that class
since it wasn't really doing much anymore after this refactoring.

The main refactoring was making it so that process assertions can be created without an
AuxiliaryProcessProxy object. The reason for this I did not want to pass an AuxiliaryProcessProxy
object that is in the middle of its destructor to another function.

* Source/WebKit/SourcesCocoa.txt:
* Source/WebKit/UIProcess/API/Cocoa/WKWebsiteDataStore.mm:
(-[WKWebsiteDataStore _forceNetworkProcessToTaskSuspendForTesting]):
* Source/WebKit/UIProcess/API/Cocoa/WKWebsiteDataStorePrivate.h:
* Source/WebKit/UIProcess/AuxiliaryProcessProxy.cpp:
(WebKit::AuxiliaryProcessProxy::~AuxiliaryProcessProxy):
* Source/WebKit/UIProcess/AuxiliaryProcessProxy.h:
* Source/WebKit/UIProcess/Cocoa/AuxiliaryProcessProxyCocoa.mm:
(WebKit::AuxiliaryProcessProxy::platformStartConnectionTerminationWatchdog):
* Source/WebKit/UIProcess/Cocoa/ProcessAssertionCocoa.mm:
(WebKit::ProcessAssertion::ProcessAssertion):
(WebKit::ProcessAndUIAssertion::ProcessAndUIAssertion):
* Source/WebKit/UIProcess/Cocoa/XPCConnectionTerminationWatchdog.h: Removed.
* Source/WebKit/UIProcess/Cocoa/XPCConnectionTerminationWatchdog.mm: Removed.
* Source/WebKit/UIProcess/ProcessAssertion.cpp:
(WebKit::ProcessAssertion::acquireAssertion):
(WebKit::ProcessAssertion::create):
(WebKit::ProcessAndUIAssertion::create):
(WebKit::ProcessAssertion::ProcessAssertion):
(WebKit::ProcessAndUIAssertion::ProcessAndUIAssertion):
(WebKit::ProcessAssertion::aquireAssertion): Deleted.
* Source/WebKit/UIProcess/ProcessAssertion.h:
* Source/WebKit/UIProcess/mac/WKTextAnimationManager.mm:
* Source/WebKit/WebKit.xcodeproj/project.pbxproj:
* Tools/TestWebKitAPI/Tests/WebKitCocoa/NetworkProcess.mm:
(isTaskSuspended):
(TEST(NetworkProcess, TerminateWhenNetworkProcessIsSuspended)):
* Tools/TestWebKitAPI/Utilities.h:
(TestWebKitAPI::Util::waitFor):

Canonical link: https://commits.webkit.org/289075@main